requirements.txt
# Core dependencies
torch==2.4.1
pytorch-lightning==2.1.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Qlib for financial data
pyqlib==0.9.6.99

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and visualization
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data processing
tqdm>=4.65.0
pyyaml>=6.0
pickle5>=0.0.12

# Additional utilities
einops>=0.7.0
yfinance>=0.2.28  # For fetching market data
statsmodels>=0.14.0  # For statistical tests

# Optional: For development and testing
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: For Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0